// back-end/src/models/interface/IBodywork.ts
export type IBodywork = {
  id?: number
  bajos_del_paragolpe_delantero: number
  capó: number
  guardabarro_delantero_derecho: number
  espejo_retrovisor_izquierdo: number
  panel_exterior___puerta_delantera_izquierda___: number
  bisagra___puerta_delantera_izquierda___: number
  manija_exterior___puerta_delantera_derecha___: number
  parabrisas_delantero: number
  paragolpe_delantero: number
  guardabarro_delantero_izquierdo: number
  parabrisas: number
  espejo_retrovisor_derecho: number
  manija___puerta_delantera_izquierda___: number
  panel_exterior___puerta_delantera_derecha___: number
  bisagra___puerta_delantera_derecha___: number
  talonera_izquierda: number
  car_id: number
}